updated seeds to include complete_internship
